Keep all special characters in boolean abstraction
Replacing special characters in variable names is mostly necessary to
adhere to the grammar supported by the FeatureIDE parser for `Node`s.
However, we already do our own parsing using Antlr4 so we can skip this
step by building the `Node` tree our selves. This also removes one more
source of bugs because overall there is less code involved.

Beware that this introduces a semantic change: The new JPP parser
abstracts `!defined(property)` to a negative literal `defined(property)`
instead of a positive literal with a prefix denoting the negation.