Inline and remove AbstractingFormulaExtractor
This class is weirdly coupled with its subclasses. It uses a regex and
assumes that some arbitrary groups exist and have the correct meaning.
Furthermore, the exception handling handles
UncheckedUnParseableFormulaException specially which should also be an
implementation detail of the code that actually throws it. Last but not
least, it only saves redundant exception handling code (which is
dependent on the subclass) as the matching code has to be repeated by
CPPDiffLineFormulaExtractor anyways (also degrading performance as the
regex needs to be matched twice for no reason).