CI: Remove the caching mechanism
After testing (everything works as excepted) and measuring the cache, we
noticed that the time improvement is not worth the additional
complexity. Hence, we decided to remove this cache but leave it in the
commit history in case it becomes relevant in the future.

This effectively reverts commit da11a13c1451f5a52672fe494ac6e20116346865
"CI: Cache dependencies between runs".