refactor: Make GitDiffer stateless
The `GitDiffer` class doesn't own (in the sense that the actual truth of
these members lies in the `Repository` class) any of its members.
Furthermore, it just bundles the same members as `Repository`, just
leaving those out, that are unneeded. The constructor of `GitDiffer`
even extracts these members from `Repository` and doesn't even allow
them to be past any other way. Furthermore, most of the interesting
functions are already static anyways.

A different approach would be to move all of methods into `Repository`.
However, there seems to be no reason to override any of the actual
diffing methods and the drawback would be a huge `Repository` class
which would serve a second purpose.