fix: Manually move generated javadoc to docs/javadoc
The dependency update of the maven-javadoc-plugin in
afd65e1d9079b9d1ea698c9ffb9e3322914e10c5 removed the possibility to
define an alternative output directory for the generated javadoc which
resulted in a warning that I missed until now. This regression was
intentional and is declared as won't fix by upstream. See
https://github.com/apache/maven-javadoc-plugin/issues/1194